perf(analysis): vectorize PSNR SSD, optimize SSIM block stats & Laplacian noise, harden PixelInspector bounds and file limits
- SIMD-accelerated quality metrics in core/analysis/AnalysisEngine_metrics.cpp:
  - AVX2/SSE2 vectorized PSNR sum-of-squared differences (SSD) for Grayscale8, RGB24, BGR24
  - SSE2 4-pixel/iter RGBA32/BGRA32 RGB-difference kernel with SIMD alpha masking
  - Zero-allocation cross-format fast paths (RGB24 vs BGR24 and RGBA32 vs BGRA32)
  - SSE2 SSIM 8x8 block inner statistics vectorization (~4x speedup)
  - SSE2 Laplacian convolution for noise estimation (16 pixels/iter)
- Bounded and hardened PixelInspector in core/analysis/PixelInspector_adjust.cpp:
  - Fixed 32-bit signed long overflow on MSVC in neighborhoodStats (switched to int64_t)
  - Hardened coordinate boundary clamping against out-of-bounds ROI inputs
- Hardened ExportReport.cpp:
  - Added null/empty buffer and non-contiguous padded stride validation in summarizeDiff
- Fixed signedness comparison in ImageAdjust.h against signed ptrdiff_t stride
- ADR-014 compliance splits and complexity gate conformance:
  - Split PixelInspector.cpp (233 lines) and PixelInspector_adjust.cpp (566 lines)
  - Split AnalysisEngine.cpp (213 lines) and AnalysisEngine_metrics.cpp (599 lines)
  - Kept all TUs strictly <= 800 lines and function bodies <= 120 lines
- Unit test suite:
  - Registered test_analysisengine / analysisengine_tests covering SIMD correctness, mathematical reference values, cross-format comparisons, SSIM, and noise estimation
- Bumped version to 1.0.56 in CMakeLists.txt, STATUS.md, CHANGELOG.md, and ADR-014

+++ src/core/image/ImageAdjust.h
@@ -41,9 +41,8 @@ inline ImageData adjustBrightness(const ImageData &src, int offset)
         lut[static_cast<size_t>(i)] = static_cast<uint8_t>(std::clamp(i + adj, 0, 255));
     }
 
-    const bool isContiguous =
-        (v.stride() == static_cast<size_t>(v.width) * static_cast<size_t>(cpp)) &&
-        (d.stride() == static_cast<size_t>(d.width) * static_cast<size_t>(cpp));
+    const bool isContiguous = (v.stride() == static_cast<ptrdiff_t>(v.width) * cpp) &&
+                              (d.stride() == static_cast<ptrdiff_t>(d.width) * cpp);
 
     if (isContiguous && cpp == ch)
     {
@@ -103,9 +102,8 @@ inline ImageData adjustContrast(const ImageData &src, float factor)
             static_cast<uint8_t>(std::clamp(static_cast<int>(std::lroundf(val)), 0, 255));
     }
 
-    const bool isContiguous =
-        (v.stride() == static_cast<size_t>(v.width) * static_cast<size_t>(cpp)) &&
-        (d.stride() == static_cast<size_t>(d.width) * static_cast<size_t>(cpp));
+    const bool isContiguous = (v.stride() == static_cast<ptrdiff_t>(v.width) * cpp) &&
+                              (d.stride() == static_cast<ptrdiff_t>(d.width) * cpp);
 
     if (isContiguous && cpp == ch)
     {
@@ -167,9 +165,8 @@ inline ImageData adjustGamma(const ImageData &src, float gamma)
             std::clamp(static_cast<int>(std::lroundf(corrected * 255.0f)), 0, 255));
     }
 
-    const bool isContiguous =
-        (v.stride() == static_cast<size_t>(v.width) * static_cast<size_t>(cpp)) &&
-        (d.stride() == static_cast<size_t>(d.width) * static_cast<size_t>(cpp));
+    const bool isContiguous = (v.stride() == static_cast<ptrdiff_t>(v.width) * cpp) &&
+                              (d.stride() == static_cast<ptrdiff_t>(d.width) * cpp);
 
     if (isContiguous && cpp == ch)
     {
@@ -238,9 +235,8 @@ inline ImageData adjustWhiteBalance(const ImageData &src, float rGain, float bGa
         lutB[static_cast<size_t>(i)] = static_cast<uint8_t>(std::clamp(bVal, 0, 255));
     }
 
-    const bool isContiguous =
-        (v.stride() == static_cast<size_t>(v.width) * static_cast<size_t>(cpp)) &&
-        (d.stride() == static_cast<size_t>(d.width) * static_cast<size_t>(cpp));
+    const bool isContiguous = (v.stride() == static_cast<ptrdiff_t>(v.width) * cpp) &&
+                              (d.stride() == static_cast<ptrdiff_t>(d.width) * cpp);
 
     if (isContiguous)
     {
